refactor: move schemas directory
