+++ src/schemas/author.ts
@@ -1,5 +1,5 @@
 import { z } from "astro:content";
-import { imageSchema } from "./_image";
+import { imageSchema } from "./image";
 
 export const authorSchema = z.object({
   name: z.string().max(25, { message: 'O nome do autor deve ter no máximo 25 caracteres' }),
